Remove copy constructor for tensor classes. Here's why:
Copy constructor is invoked implicitly when passing a tensor to a kernel. Copy constructor was coded
to duplicate the T*data arrays and all results were written to duplicates. Result was that on the outside, first kernel did
nothing at all. This was 'fixed' by removing the copy-constructor. In this kernel, copying will not be tolerated.