Fix incorrect allocation of cuda memory.
There was a subtle malloc issue with bas.cgto array. First, bas is a plain struct which contains 2d tensor of cgtos, which themselves are plain const structs with predefined shape.
The problem was that the simple `to_device()` function simply copied over the tensor of cgtos, without copying over the underlying array of cgtos. Inside the kernel,
we use plain structures, like `const basis_type bas`. this type contains, as a member a  `tensor2d_t<cgto_type>`, which contains `dim1, dim2, and cgto_type* data`. This past part
(data) was somehow accessible *from inside* the kernel, without me explicitly transferring it to the device. This meant that the first kernel (intra-atomic) worked correctly. The second one
got a "corrupted" version of `bas.cgto`. This was fixed by `const`-ifying most tensors and making sure the memory transfer was correctly done.