Fix lx not being correctly generated and fix s3d accumulation.
LX was not correctly generated by AI. The shape was not correctly interpreted and the result was an incorrect placement of values in the final
hamiltonian. Additionally, s3d accumulation, done by transform0 was not correct. `overlap` variable is a pre-allocated maximum scratch size for the integral. It will
be larger than or equal to size than s3d. When copying data to `overlap`, iteration must happen on smaller array `s3d`. This was fixed.